Example where we always cache state
So here's an example where we always cache the state. NOW we run into
caching problem, though _internally_ it's easier to use because we
always have the most up to date state.

Frankly I think this way is _more_ error prone, not the least of which
is that you need to use the wrapper struct _everywhere_ instead of just
internally...