Added IPC module, added nice inner_try!
Stupid globals, thank goodness for my macros.

The ipc stuff still needs to be fleshed out, e.g I'm not sure of the
exact mechanism yet I'm going to use to send things between.

For this week though, I'm going to use a shared mutex to send the Cairo
buffer since I just want the top bar to render because that would be
cool.

In the future though, it will not do something so terrible.